% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlearn.R
\name{xlearn}
\alias{xlearn}
\alias{predict.xl.model}
\title{Predict method for xlearn model}
\usage{
xlearn(params = list(), data, label, type = c("linear", "fm", "ffm"),
  validate = NULL, model.path = "./model.out")

\method{predict}{xl.model}(object, newdata, out.path = "./pred.out")
}
\description{
Predict method for xlearn model
}
